+++ src/meltano/cli/initialize.py
@@ -8,6 +8,7 @@ from meltano.core.project_init_service import (
     ProjectInitServiceError,
 )
 from meltano.core.plugin_install_service import PluginInstallService
+from meltano.core.tracking import GoogleAnalyticsTracker
 from . import cli
 
 EXTRACTORS = "extractors"
@@ -17,12 +18,20 @@ ALL = "all"
 
 @cli.command()
 @click.argument("project_name")
-def init(project_name):
+@click.option(
+    "--no_usage_stats", help="Do not send anonymous usage stats.", is_flag=True
+)
+def init(project_name, no_usage_stats):
     init_service = ProjectInitService(project_name)
     try:
         project = init_service.init()
         init_service.echo_instructions()
+
+        tracker = GoogleAnalyticsTracker(project)
+        if no_usage_stats:
+            tracker.update_permission_to_track(False)
+        else:
+            tracker.track_meltano_init(project_name=project_name)
     except ProjectInitServiceError as e:
-        print(e)
         click.secho(f"Directory {project_name} already exists!", fg="red")
         raise click.Abort()